refactor: Interface 분리
