+++ src/main/kotlin/baseball/domain/numberRules/ComputerGenerateNumberRule.kt
@@ -1,0 +1,5 @@
+package baseball.domain.numberRules
+
+interface ComputerGenerateNumberRule {
+    fun generateRandomNumber(): MutableList<Int>
+}
