+++ src/main/kotlin/baseball/domain/numberRules/PlayerSelectNumberRule.kt
@@ -1,0 +1,5 @@
+package baseball.domain.numberRules
+
+interface PlayerSelectNumberRule {
+    fun selectNumber(): List<Int>
+}
